+++ changelog.md
@@ -7,6 +7,7 @@
 - Added/changed custom UI logic, now based on ImGui
 - Added a button to open/close the menu and the plugin management menu itself
 - Added basic templates for UI widgets
+- Added configuration system for the plugin
 - Fix `Zip Slip` vulnerability in `JarTools` extraction method (#9)
 - Fixed (separated) logic for loading plugins from the server and client sides